requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
Pillow>=8.0.0
scipy>=1.7.0
scikit-image>=0.18.0
tqdm>=4.60.0